More X11 cruft
Need to package libxcb and wire Cairo and libgdiplus up to it for
Windows Forms to possibly work.

Packaged the X.org libXau, because while AIX ships a libXau, it
appears that PASE does not.

Some compatibility stubs and M4 macros have been packaged - I note
that none of the AIX packagers had mentioned or packaged xorg
macro package, so that was annoying to figure out! pthread stubs
is a no-op package basically since 0.4, but I include it anyways
instead of vendoring it into an SRPM.